[ADD] set content of controller spreasheet class

+++ spreadsheet/src/main/java/edu/upc/etsetb/arqsoft/spreadsheet/entities/DefaultSpreadsheetFactory.java
@@ -25,7 +25,7 @@ public class DefaultSpreadsheetFactory extends SpreadsheetFactory {
     }
 
     @Override
-    public Number createNumber(Float number) {
+    public Number createNumber(Double number) {
         throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
     }
